fixed move grid line mode (double-click or right-click to delete)
git-svn-id: svn://136.177.114.72/svn_GW/WPhast/trunk@5546 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/GridActor.h
@@ -176,7 +176,7 @@ protected:
 	int                   AxisIndex;
 	int                   PlaneIndex;
 	double                CurrentPoint[3];
-	std::map<float, int>  ValueToIndex[3];
+	std::map<double, int> ValueToIndex[3];
 
 
 	CGridKeyword        m_gridKeyword;
